[issue-158] Added pulling images to parent script
